bear: Changed the Stream detection method on calls to the StreamInfo::GetFindServiceNameCur() now includes channel number in the test. Added new Media Formats to make the demux pins the same as the TSFileSource media types. Also changed the sink demux types as well. Removed the file build delay, loading of the TSFileSource is now instant, although the settings for file build method can be still valid if the values are not zero. Added in AddDemuxpins method on the TSFileSource class method prior to connection. Added in Forced Pin mode to FileSource::Load call. see TSFileSource notes.
TSFileSource Changes:
Added some critical locks to the demux, pidinfoarray & stream parser classes.
Changed the duration update method to only parse if we have the pids parsed.
Improved the startup delay by using a cold start method, the fillbuffer does not parse for pcr's until the pids are parsed. This means that the graph clock will not be updated until the pids are parsed. This occurs when there is sufficient file size to correctly parse the data.
Improved the demux::cleardemuxpin() method byreplacing some calls with a vetetdemuxpin() method so it doesn't clear the pid if the new pid is the same value. 
Added in a reference clock delay when we reach the end of the file, will slightly improve as time progresses to stay ahead.
Added a RefreshTSBufferFile() call to the MultiFileReader::GetFileSize() function.
Added a force output pin type method on the FileSource::load() function, this allows the graph to build with a program pin if required. i.e. If pmt param is media subtype program else Transport pin. If pmt is null then pin type is derived from the file parsing.


git-svn-id: http://svn.nate.deepcreek.org.au/svn/DigitalWatch/trunk@119 a123065e-f578-804e-ac06-b1832d43b128